requirements.txt
# Core dependencies
numpy
scipy
pandas
matplotlib
plotly
dash
dash-bootstrap-components>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0

# Data processing
scikit-learn>=1.0.0
librosa>=0.8.0

# Database
#sqlite3

# Serial communication
pyserial

# MQTT
paho-mqtt

# Audio processing
#sounddevice>=0.4.0
#soundfile>=0.10.0

# Threading and multiprocessing
#threading
#multiprocessing

# Utilities
#python-dateutil>=2.8.0
#pytz>=2021.1
#requests>=2.25.0

# Development and testing
#pytest>=6.0.0
#pytest-cov>=2.10.0
#black>=21.0.0
#flake8>=3.8.0

# Documentation
#sphinx>=4.0.0
#4sphinx-rtd-theme>=0.5.0